feat: Added search page
This change introduces a new page to allow searching of books.

The hierarcy of componnents has been update to support reuse of logic
such as book shelf updates. (DRY)

A new debouce helper function has been added to reduce the load on the
server as the seach functionality update as the user types